+++ maplibre_gl_platform_interface/lib/src/method_channel_maplibre_gl.dart
@@ -136,6 +136,7 @@ class MapLibreMethodChannel extends MapLibrePlatform {
     Set<Factory<OneSequenceGestureRecognizer>>? gestureRecognizers,
   ) {
     if (defaultTargetPlatform == TargetPlatform.android) {
+      creationParams['options']?['useHybridComposition'] = useHybridComposition;
       if (useHybridComposition) {
         return PlatformViewLink(
           viewType: 'plugins.flutter.io/maplibre_gl',